Joeurs Platton 0.8  run command

+++ src/jeu/patrouille/coeur/actions/BandageAction.java
@@ -21,7 +21,7 @@ public class BandageAction extends BaseAction{
 
     @Override
     public void calculeActionPointDesAction() throws Exception {
-        tempActivite= type.TN();
+        tempActivite=(int) type.TN();
     }
 
 }
